refactor: address PR review feedback
Addresses Kevin's review comments:

1. Move serialization functions to dedicated utility file
   - Created src/serialization.ts with serializeLoadSubsetOptions,
     serializeExpression, and serializeValue functions
   - Keeps query.ts focused on query logic

2. Fix return type and use undefined instead of null
   - Changed serializeLoadSubsetOptions return type from `unknown` to
     `string | undefined`
   - Returns undefined instead of null for empty options
   - Updated usage to conditionally append only when serialized result
     is not undefined

3. Add missing CompareOptions properties to orderBy serialization
   - Now includes stringSort, locale, and localeOptions properties
   - Properly handles the StringCollationConfig union type with
     conditional serialization for locale-specific options

All runtime tests pass (65/65 in query.test.ts).
