+++ packages/query-db-collection/src/query.ts
@@ -7,12 +7,12 @@ import {
   QueryKeyRequiredError,
 } from "./errors"
 import { createWriteUtils } from "./manual-sync"
+import { serializeLoadSubsetOptions } from "./serialization"
 import type {
   BaseCollectionConfig,
   ChangeMessage,
   CollectionConfig,
   DeleteMutationFnParams,
-  IR,
   InsertMutationFnParams,
   LoadSubsetOptions,
   SyncConfig,
@@ -303,119 +303,6 @@ class QueryCollectionUtilsImpl {
       (observer) => observer.getCurrentResult().fetchStatus
     )
   }
-}
-
-/**
- * Serializes LoadSubsetOptions into a stable, hashable format for query keys
- * @internal
- */
-function serializeLoadSubsetOptions(
-  options: LoadSubsetOptions | undefined
-): unknown {
-  if (!options) {
-    return null
-  }
-
-  const result: Record<string, unknown> = {}
-
-  if (options.where) {
-    result.where = serializeExpression(options.where)
-  }
-
-  if (options.orderBy?.length) {
-    result.orderBy = options.orderBy.map((clause) => ({
-      expression: serializeExpression(clause.expression),
-      direction: clause.compareOptions.direction,
-      nulls: clause.compareOptions.nulls,
-    }))
-  }
-
-  if (options.limit !== undefined) {
-    result.limit = options.limit
-  }
-
-  return JSON.stringify(Object.keys(result).length === 0 ? null : result)
-}
-
-/**
- * Recursively serializes an IR expression for stable hashing
- * @internal
- */
-function serializeExpression(expr: IR.BasicExpression | undefined): unknown {
-  if (!expr) {
-    return null
-  }
-
-  switch (expr.type) {
-    case `val`:
-      return {
-        type: `val`,
-        value: serializeValue(expr.value),
-      }
-    case `ref`:
-      return {
-        type: `ref`,
-        path: [...expr.path],
-      }
-    case `func`:
-      return {
-        type: `func`,
-        name: expr.name,
-        args: expr.args.map((arg) => serializeExpression(arg)),
-      }
-    default:
-      return null
-  }
-}
-
-/**
- * Serializes special JavaScript values (undefined, NaN, Infinity, Date)
- * @internal
- */
-function serializeValue(value: unknown): unknown {
-  if (value === undefined) {
-    return { __type: `undefined` }
-  }
-
-  if (typeof value === `number`) {
-    if (Number.isNaN(value)) {
-      return { __type: `nan` }
-    }
-    if (value === Number.POSITIVE_INFINITY) {
-      return { __type: `infinity`, sign: 1 }
-    }
-    if (value === Number.NEGATIVE_INFINITY) {
-      return { __type: `infinity`, sign: -1 }
-    }
-  }
-
-  if (
-    value === null ||
-    typeof value === `string` ||
-    typeof value === `number` ||
-    typeof value === `boolean`
-  ) {
-    return value
-  }
-
-  if (value instanceof Date) {
-    return { __type: `date`, value: value.toJSON() }
-  }
-
-  if (Array.isArray(value)) {
-    return value.map((item) => serializeValue(item))
-  }
-
-  if (typeof value === `object`) {
-    return Object.fromEntries(
-      Object.entries(value as Record<string, unknown>).map(([key, val]) => [
-        key,
-        serializeValue(val),
-      ])
-    )
-  }
-
-  return value
 }
 
 /**
@@ -748,7 +635,8 @@ export function queryCollectionOptions(
       } else if (syncMode === `on-demand`) {
         // Static queryKey in on-demand mode: automatically append serialized predicates
         // to create separate cache entries for different predicate combinations
-        key = [...queryKey, serializeLoadSubsetOptions(opts)]
+        const serialized = serializeLoadSubsetOptions(opts)
+        key = serialized !== undefined ? [...queryKey, serialized] : queryKey
       } else {
         // Static queryKey in eager mode: use as-is
         key = queryKey
